refactor: bit of tidy up, split out into multiple svelte files
